requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
future==0.18.2
idna==2.8
image==1.5.27
Pillow==7.0.0
pyglet==1.4.9
pytesseract==0.3.1
pyttsx3==2.71
pytz==2019.3
requests==2.22.0
RPi.GPIO==0.7.0
sqlparse==0.3.0
urllib3==1.25.7
